# scanforge-security-platform/apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScanForge API",
    "description": "Repository security scanning and vulnerability management platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/internal/scans/{scan_id}/artifacts/upload-url": {
      "post": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Create Artifact Upload Url",
        "operationId": "create_artifact_upload_url_api_v1_internal_scans__scan_id__artifacts_upload_url_post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/notifications": {
      "post": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Create Notification",
        "operationId": "create_notification_api_v1_internal_notifications_post",
        "parameters": [
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/scans/{scan_id}/complete": {
      "post": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Complete Scan",
        "operationId": "complete_scan_api_v1_internal_scans__scan_id__complete_post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/scans/{scan_id}/status": {
      "patch": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Update Scan Status Internal",
        "operationId": "update_scan_status_internal_api_v1_internal_scans__scan_id__status_patch",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanProgressUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/scans/{scan_id}/scanner-runs": {
      "post": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Create Scanner Run",
        "operationId": "create_scanner_run_api_v1_internal_scans__scan_id__scanner_runs_post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScannerRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/scanner-runs/{run_id}": {
      "patch": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Update Scanner Run",
        "operationId": "update_scanner_run_api_v1_internal_scanner_runs__run_id__patch",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScannerRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/scans/{scan_id}/findings": {
      "post": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Persist Scan Findings",
        "operationId": "persist_scan_findings_api_v1_internal_scans__scan_id__findings_post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersistFindingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/repositories/{repo_id}/clone-url": {
      "get": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Get Repository Clone Url",
        "description": "Return an authenticated clone URL for the worker to use.",
        "operationId": "get_repository_clone_url_api_v1_internal_repositories__repo_id__clone_url_get",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repo Id"
            }
          },
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/scans/{scan_id}/execution-context": {
      "get": {
        "tags": [
          "internal",
          "internal"
        ],
        "summary": "Get Scan Execution Context",
        "operationId": "get_scan_execution_context_api_v1_internal_scans__scan_id__execution_context_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "X-Worker-Credential",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Credential"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactUploadRequest": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "execution_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Execution Revision"
          },
          "scanner_name": {
            "type": "string",
            "title": "Scanner Name"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "application/json"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "execution_revision",
          "scanner_name",
          "filename",
          "size_bytes"
        ],
        "title": "ArtifactUploadRequest"
      },
...
      "CreateScannerRunRequest": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "execution_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Execution Revision"
          },
          "scanner_name": {
            "type": "string",
            "title": "Scanner Name"
          },
          "scanner_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanner Version"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "execution_revision",
          "scanner_name"
        ],
        "title": "CreateScannerRunRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}